Psuedocode written and validMoves function working.
Some starting pseudocode for the basic functions.

validMove() function complete that takes the current location and goal then finds the valid moves from current and adds them all to an array.

Before add to array check if match with goal and if the new move matches return true instead of moves array.